+++ models/model_events_py/dJzdsigmaz.cpp
@@ -6,7 +6,7 @@
 namespace amici {
 namespace model_model_events_py {
 
-void dJzdsigma_model_events_py(realtype *dJzdsigma, const int iz, const realtype *p, const realtype *k, const realtype *z, const realtype *sigmaz, const realtype *mz){
+void dJzdsigmaz_model_events_py(realtype *dJzdsigmaz, const int iz, const realtype *p, const realtype *k, const realtype *z, const realtype *sigmaz, const realtype *mz){
     const realtype z1_ = z[0];
     const realtype z2_ = z[1];
     const realtype sigma_z1_ = sigmaz[0];
@@ -16,10 +16,10 @@ void dJzdsigma_model_events_py(realtype *dJzdsigma, const int iz, const realtype
 
     switch(iz) {
         case 0:
-            dJzdsigma[0] = 1.0/sigma_z1_ - 1.0*std::pow(-mz1_ + z1_, 2)/std::pow(sigma_z1_, 3);
+            dJzdsigmaz[0] = 1.0/sigma_z1_ - 1.0*std::pow(-mz1_ + z1_, 2)/std::pow(sigma_z1_, 3);
             break;
         case 1:
-            dJzdsigma[1] = 1.0/sigma_z2_ - 1.0*std::pow(-mz2_ + z2_, 2)/std::pow(sigma_z2_, 3);
+            dJzdsigmaz[1] = 1.0/sigma_z2_ - 1.0*std::pow(-mz2_ + z2_, 2)/std::pow(sigma_z2_, 3);
             break;
     }
 }