+++ models/model_steadystate_py/dJydsigmay.cpp
@@ -4,9 +4,9 @@
 #include <algorithm>
 
 namespace amici {
-namespace model_model_robertson_py {
+namespace model_model_steadystate_py {
 
-void dJydsigma_model_robertson_py(realtype *dJydsigma, const int iy, const realtype *p, const realtype *k, const realtype *y, const realtype *sigmay, const realtype *my){
+void dJydsigmay_model_steadystate_py(realtype *dJydsigmay, const int iy, const realtype *p, const realtype *k, const realtype *y, const realtype *sigmay, const realtype *my){
     const realtype obs_x1_ = y[0];
     const realtype obs_x2_ = y[1];
     const realtype obs_x3_ = y[2];
@@ -19,16 +19,16 @@ void dJydsigma_model_robertson_py(realtype *dJydsigma, const int iy, const realt
 
     switch(iy) {
         case 0:
-            dJydsigma[0] = 1.0/sigma_obs_x1_ - 1.0*std::pow(-mobs_x1_ + obs_x1_, 2)/std::pow(sigma_obs_x1_, 3);
+            dJydsigmay[0] = 1.0/sigma_obs_x1_ - 1.0*std::pow(-mobs_x1_ + obs_x1_, 2)/std::pow(sigma_obs_x1_, 3);
             break;
         case 1:
-            dJydsigma[1] = 1.0/sigma_obs_x2_ - 1.0*std::pow(-mobs_x2_ + obs_x2_, 2)/std::pow(sigma_obs_x2_, 3);
+            dJydsigmay[1] = 1.0/sigma_obs_x2_ - 1.0*std::pow(-mobs_x2_ + obs_x2_, 2)/std::pow(sigma_obs_x2_, 3);
             break;
         case 2:
-            dJydsigma[2] = 1.0/sigma_obs_x3_ - 1.0*std::pow(-mobs_x3_ + obs_x3_, 2)/std::pow(sigma_obs_x3_, 3);
+            dJydsigmay[2] = 1.0/sigma_obs_x3_ - 1.0*std::pow(-mobs_x3_ + obs_x3_, 2)/std::pow(sigma_obs_x3_, 3);
             break;
     }
 }
 
-} // namespace model_model_robertson_py
+} // namespace model_model_steadystate_py
 } // namespace amici